Eliminate implicit conversions
Use explicit conversions everywhere. Adding bounds check assertions when
necessary.

+++ src/vsync.c
@@ -38,7 +38,7 @@ static int vsync_drm_wait(session_t *ps) {
 
 	do {
 		ret = ioctl(ps->drm_fd, DRM_IOCTL_WAIT_VBLANK, &vbl);
-		vbl.request.type &= ~_DRM_VBLANK_RELATIVE;
+		vbl.request.type &= ~(uint)_DRM_VBLANK_RELATIVE;
 	} while (ret && errno == EINTR);
 
 	if (ret)
@@ -90,9 +90,9 @@ static bool vsync_opengl_oml_init(session_t *ps) {
 	return glxext.has_GLX_OML_sync_control;
 }
 
-static inline bool vsync_opengl_swc_swap_interval(session_t *ps, unsigned int interval) {
+static inline bool vsync_opengl_swc_swap_interval(session_t *ps, int interval) {
 	if (glxext.has_GLX_MESA_swap_control)
-		return glXSwapIntervalMESA(interval) == 0;
+		return glXSwapIntervalMESA((uint)interval) == 0;
 	else if (glxext.has_GLX_SGI_swap_control)
 		return glXSwapIntervalSGI(interval) == 0;
 	else if (glxext.has_GLX_EXT_swap_control) {
